Changed some permissions and cleaned up

+++ installers/nurc_full_installer.sh
@@ -1,31 +1,26 @@
 opencv_latest=false
-
-if [ ! -e ~/.bash_profile ]; then
-    touch ~/.bash_profile
-    echo -e "if [ -f ~/.bashrc ]; then\n    source ~/.bashrc\nfi" >> ~/.bash_profile
-else
-    if ! grep -q "source ~/.bashrc" ~/.bashrc; then
-        echo -e "if [ -f ~/.bashrc ]; then\n    source ~/.bashrc\nfi" >> ~/.bash_profile
-    fi
-fi
+ros_install=true
 
 for i in "$@"; do
     if [ $i == '--opencv' ] && [ "$opencv_latest" = false ]; then
         echo "Queuing the installer for the OpenCV 2 latest bundle."
-        opencv_latest=true
-    fi
+				opencv_latest=true
+    elif [ $i == '--no-ros' ]; then
+				ros_install=false
+		fi
 done
 
 sudo apt-get update
 sudo apt-get purge -y firefox thunderbird* pidgin* link-grammar-dictionaries-en
 sudo apt-get install -y chromium-browser
-sudo apt-get install -y flashplugin-installer
+sudo apt-get install -y flashplugin-installer pepperflashplugin-nonfree
+sudo apt-get install -y vim byobu
 sudo apt-get install -y git
 sudo apt-get install -y build-essential
 
 ## Set up workspace
-mkdir -p ~/Code/nurobotics_ws
-mkdir -p ~/Code/ros_ws
+sudo -u $USER mkdir -p ~/Code/nurobotics_ws
+sudo -u $USER mkdir -p ~/Code/ros_ws
 # git clone https://github.com/mjzh/Hermes.git ~/Code/nurobotics_ws
 # git clone https://github.com/NURobotics/lacrosse_bot.git ~/Code/nurobotics_ws
 # git clone https://github.com/NURobotics/darwin-op.git ~/Code/nurobotics_ws
@@ -33,14 +28,19 @@ mkdir -p ~/Code/ros_ws
 source dc_workshops_installer.sh
 source hermes_installer.sh
 source darwin_installer.sh
-source ros_indigo_installer.sh
+
+if [ "$ros_install" = true ]; then
+		source ros_indigo_installer.sh
+fi
 
 ## OpenCV Installation
-if [ "$opencv_latest" = true ]; then
+if [ "$opencv_latest" = true ] && [ "$ros_install" = false ]; then
     git clone https://github.com/jayrambhia/Install-OpenCV.git /tmp
     sudo chmod +x /tmp/Install-OpenCV/Ubuntu/opencv_latest.sh
     sudo sh /tmp/Install-OpenCV/Ubuntu/opencv_latest.sh
     sudo rm -rf /tmp/Install-OpenCV
+elif [ "$opencv_latest" = true ]; then
+		echo "OpenCV was installed as part of the ROS package."
 fi
 
 ## Configure Look and Feel
